+++ lib/hi_energy_ai/resources/tags.rb
@@ -7,8 +7,6 @@ module HiEnergyAi
         get("/tags", params: params)
       end
 
-      alias search list
-
       def advertisers(id, **params)
         get("/tags/#{id}/advertisers", params: params)
       end